Add pint.json and configure code style rules
Added a new pint.json file to determine the set of rules for code style checking. This is necessary to enforce code consistency across the project. The 'preset' option is set to 'laravel' and specific rules are explicitly defined. This includes options for how to handle binary operator spaces, order of imports, phpdoc formatting, method casing, braces positioning, etc. This configuration ensures that all the code following these rules is neat, readable, and follows best practices.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>