Remove php-cs-fixer-config.php and php-cs-fixer-rules.php
Removed php-cs-fixer-config.php and php-cs-fixer-rules.php files. This decision was made due to these files becoming redundant as all the necessary configurations and rules are now set up directly on the pint.json file. This provides a streamlined and consolidated way of managing code conventions and rules for the project.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>