Update phpstan.neon.dist configuration
The configuration file for PHPStan, phpstan.neon.dist, has been updated for stricter type checking and error reporting. The "checkMissingIterableValueType" and "reportUnmatchedIgnoredErrors" were set to true, enabling detection of missing iterable value types and unmatched ignored errors respectively. Also, the "checkGenericClassInNonGenericObjectType" was added and set to false, so as not to report an error whenever a generic class is used in a non-generic object type in the application.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>