Adjust GitHub Actions to track both main and master branches
Added master branch to the list of branches to be tracked by GitHub Actions in addition to the main branch. This ensures that GitHub workflows are triggered on push or pull request events happening not just on the main branch but also on the master branch, thus covering more potential workflows in the project's lifecycle.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>