Simplify GitHub Actions Workflow
The Pest workflow has been simplified by removing matrix strategy and replacing complex dependency installation with `ramsey/composer-install` action. The changes result in a cleaner workflow file that is easier to maintain and understand. It also simplifies the job configuration by running it only on `ubuntu-latest`, thus reducing potential issues related to differences in OS. This change supports the principle of maintaining simplicity and clarity in our codebase.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>