Update Laravel Pint Action in GitHub workflow
Added configurations for preset and configPath in Laravel Pint Action to ensure adherence to our code styling conventions. The preset is set as Laravel and configPath for pint.json is specified to streamline and standardize the PHP code style fixing process. This preventive approach helps reduce time spent on code review and enhances code readability.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>