Modify Infection workflow conditions in GitHub
Adjusted the event conditions for the Infection GitHub workflow. Previously, it only ran on workflow calls and push events. Now, it will also trigger on Pull Request events to ensure that incoming code changes do not introduce any infections. The change aims to enhance the workflow efficiency and code quality.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>