Refactor GitHub Actions workflow conditions
Updated the pull and push event conditions for PHPstan, Pint, and Infection workflows. Previously, they were grouped under a common reference, but have now been defined separately for each event. The paths for workflows have also been specified. This has been done to improve the readability of the code and to facilitate future edits to the workflows.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>