Update project type and dependencies in composer.json
Changed the package type from "library" to "composer-plugin" in composer.json and added new dependencies while setting up plugin infrastructure. The "composer-plugin-api" was added to "require", "composer/composer" was moved to "require-dev". Moreover, a PSR-4 autoload mapping for the "Phonyland\DevTools" namespace was established. This change is necessitated to set up the project as a composer plugin and to fulfill its demand for specific dependencies.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>