Add PSR-4 autoload to composer.json
Added an "autoload" section to composer.json to enable PSR-4 autoloading for the 'Phonyland\\DevTools' namespace pointing to the 'src/' directory. This change improves the usage of classes within the namespace by not requiring manual addition of each class file. It also streamlines the code organization and adheres to PSR-4 autoloading standards.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>