Add LintCommand in CustomCommandProvider to improve execution
The LintCommand class is now included in CustomCommandProvider.php to streamline the linting process. This was done after relocating the LintCommand file to `/src/Commands`, thus necessitating this change for the correct path execution. It would improve the project's file organization, and enhance the lint command's functionality by easing its access and execution. It also includes providing new command aliases and defining a return type for configure function to maintain code clarity.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>