Improve linting process by modifying LintCommand execution
The execution line in the LintCommand class was changed to improve the linting process by correctly referencing the relocated file and easing its execution. This involved adding the LintCommand in CustomCommandProvider.php and adjusting the LintCommand execution line in the process. This ensures proper file referencing, adds new command aliases, and defines a clearer return type for configure function.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>