Add PestCommand for running unit tests
A new PestCommand class is introduced, which can be used to perform unit testing on the existing codebase using Pest. This gets rid of the old configuration and command for pest within the composer scripts, to avoid potential conflicts and also to keep composer scripts clean. We also updated the list of available commands within the CustomCommandProvider by adding the new PestCommand to the array.
This ensures a centralized handling of commands, improving maintainability.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>