Refactor PHPStan execution to a custom command
As part of improving application maintainability, PHPStan execution has been refactored from a direct composer script to a custom command within the application. The 'phpstan' composer script has been removed from composer.json. A new PhpstanCommand class has been added in the Commands directory. This class wraps the PHPStan execution logic. This new command is also added to the list of commands in theCustomCommandProvider class. The benefit of this change is a clearer and easier maintainability of commands within the application.

Signed-off-by: Yunus Emre Deligöz <ye@deligoz.me>